requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
Pillow==5.3.0
PyJWT==1.7.0
pytz==2018.7
